+++ backend/src/baserow/contrib/database/management/commands/migrate_formulas.py
@@ -1,6 +1,9 @@
 from django.core.management.base import BaseCommand
 
-from baserow.contrib.database.formula.migrations.handler import FormulaMigrationHandler
+from baserow.contrib.database.formula.migrations.handler import (
+    DEFAULT_FORMULA_MIGRATION_BATCH_SIZE,
+    FormulaMigrationHandler,
+)
 from baserow.contrib.database.table.cache import clear_generated_model_cache
 
 
@@ -20,11 +23,19 @@ class Command(BaseCommand):
             help="By default this command first clears the model cache. If you have"
             " want to disable this provide this flag.",
         )
+        parser.add_argument(
+            "--batch-size",
+            type=int,
+            default=DEFAULT_FORMULA_MIGRATION_BATCH_SIZE,
+            help="Formulas will be updated in batches of size according to this "
+            "parameter.",
+        )
 
     def handle(self, *args, **options):
         dont_clear_model_cache = options.get("dont_clear_model_cache", False)
+        batch_size = options.get("batch_size")
 
         if not dont_clear_model_cache:
             clear_generated_model_cache()
 
-        FormulaMigrationHandler.migrate_formulas_to_latest_version()
+        FormulaMigrationHandler.migrate_formulas_to_latest_version(batch_size)